workspace_events_webhook/requirements.txt
functions-framework>=3.0.0
google-cloud-aiplatform[agent_engines]>=1.112.0
google-api-python-client>=2.160.0
google-auth>=2.36.0
